Move assistant-related message to `protocol`
They are very specific to the current test suite, but since the
assistant is intended to be shared, it makes sense to have them in
`protocol` in principle.
